+++ src/PriceComparator.java
@@ -3,6 +3,6 @@ import java.util.Comparator;
 public class PriceComparator implements Comparator<Order> {
     @Override
     public int compare(Order o1, Order o2) {
-       return o1.getPrice() - o2.getPrice();
+       return (int) (o1.getPrice() - o2.getPrice());
     }
 }
